requirements.txt
django
djangorestframework
drf-spectacular

pandas
openpyxl
odfpy
